% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageDigitizer.R
\name{imageDigitizer}
\alias{imageDigitizer}
\title{R/shiny app for digitizing points in images.}
\usage{
imageDigitizer()
}
\description{
Sets up a "shiny" graphical user interface (GUI) for digitizing points in images, by
means of mouse clicks. The interface is reasonably self-explanatory.
}
\section{Developer plans}{

\enumerate{
\item Handle single-scale cases by not requiring 2 clicks on each axis.
\item Provide controls for zooming in and out.
\item Provide a way to read files created by previous uses, to extend an
analysis beyond a single session.
}
}

