% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imager.R
\name{imager}
\alias{imager}
\title{imager}
\usage{
imager()
}
\description{
A shiny graphical user interface (GUI) for digitizing points in images, by
means of mouse clicks. The GUI is meant to be reasonably self-explanatory.
}
