% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageDigitizer.R
\name{imageDigitizer}
\alias{imageDigitizer}
\title{R/shiny app for digitizing points in images.}
\usage{
imageDigitizer(fourCorners = TRUE)
}
\arguments{
\item{fourCorners}{logical value indicating (FIXME: describe this ... if we keep it!)}
}
\description{
A shiny graphical user interface (GUI) for digitizing points in images, by
means of mouse clicks. The GUI is meant to be reasonably self-explanatory.
}
